Address widget color review feedback
