+++ Sources/Shared/Widget/WidgetStateColorRule.swift
@@ -1,51 +1,16 @@
 import Foundation
 
 public struct WidgetStateColorRule: Codable, Equatable, Hashable {
-    public enum Comparison: String, Codable, CaseIterable {
-        case lessThan
-        case greaterThan
-    }
-
-    public enum Target: String, CaseIterable, Codable {
-        case state
-        case icon
-        case background
-
-        public init(from decoder: Decoder) throws {
-            let container = try decoder.singleValueContainer()
-            let value = try container.decode(String.self)
-            switch value {
-            case "text", "state":
-                // `text` was used before dynamic colors distinguished the title from the state.
-                self = .state
-            case "icon":
-                self = .icon
-            case "background":
-                self = .background
-            default:
-                throw DecodingError.dataCorruptedError(
-                    in: container,
-                    debugDescription: "Unknown widget state color target: \(value)"
-                )
-            }
-        }
-
-        public func encode(to encoder: Encoder) throws {
-            var container = encoder.singleValueContainer()
-            try container.encode(rawValue)
-        }
-    }
-
-    public let comparison: Comparison
+    public let comparison: WidgetStateColorRuleComparison
     public let threshold: Double
     public let color: String
-    public let target: Target
+    public let target: WidgetStateColorRuleTarget
 
     public init(
-        comparison: Comparison,
+        comparison: WidgetStateColorRuleComparison,
         threshold: Double,
         color: String,
-        target: Target
+        target: WidgetStateColorRuleTarget
     ) {
         self.comparison = comparison
         self.threshold = threshold
@@ -64,7 +29,7 @@ public struct WidgetStateColorRule: Codable, Equatable, Hashable {
 
     public static func matchingRule(
         in rules: [Self],
-        target: Target,
+        target: WidgetStateColorRuleTarget,
         value: Double
     ) -> Self? {
         rules.first { $0.target == target && $0.matches(value) }
