+++ Sources/Shared/Widget/WidgetStateColorRuleComparison.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum WidgetStateColorRuleComparison: String, Codable, CaseIterable {
+    case lessThan
+    case greaterThan
+}
